requirements.txt
asgiref==3.4.1
certifi==2025.4.26
charset-normalizer==2.0.12
Django==3.2.25
django-ckeditor==6.3.2
django-js-asset==2.0.0
idna==3.10
Pillow==8.4.0
Pygments==2.14.0
pytz==2025.2
requests==2.27.1
sqlparse==0.4.4
typing_extensions==4.1.1
urllib3==1.26.20
